# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Weather Forecast APIs - SkyAPI Inc.",
    "version": "1.0.0",
    "description": "Reference design documentation of Weather Forecast APIs based on the OpenAPI 3.0 specification. Created by Rohit Parida",
    "termsOfService": "https://skyweather.net/terms",
    "contact": {
      "email": "contact@skyweather.net",
      "name": "Contact SkyAPI Inc.",
      "url": "https://skyweather.net/contacts"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "servers": [
    {
      "url": "localhost:8080",
      "variables": {},
      "description": "for testing on local computer"
    }
  ],
...
  "paths": {
...
    "/v1/locations/{code}": {
      "summary": "Operations for managing a specific location identified by location code",
      "parameters": [
        {
          "name": "code",
          "in": "path",
          "schema": {
            "type": "string",
            "example": "NYC_US"
          },
          "required": true,
          "description": "code of the location"
        }
      ],
      "get": {
        "summary": "Returns details of a location",
        "description": "Clients use this API to find an existing location in the database by a specific code",
        "operationId": "getLocation",
        "responses": {
          "200": {
            "description": "Successful retrieval operation. Location found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string",
                      "description": "unique code of a location",
                      "example": "LACA_US"
                    },
                    "city_name": {
                      "type": "string",
                      "description": "name of the city",
                      "example": "Los Angeles"
                    },
                    "region_name": {
                      "type": "string",
                      "description": "name of the region, state or province",
                      "example": "California"
                    },
                    "country_code": {
                      "type": "string",
                      "description": "2 characters representing country code based on ISO 3166 standard",
                      "example": "US"
                    },
                    "country_name": {
                      "type": "string",
                      "description": "name of the country based on ISO 3166 standard",
                      "example": "United States of America"
                    },
                    "enabled": {
                      "type": "boolean",
                      "description": "whether the location enabled (true) or disabled (false)",
                      "example": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No location found with the given code."
          }
        },
        "tags": [
          "Location"
        ]
      },
      "delete": {
        "summary": "Removes an existing location found by a specific code",
        "description": "Clients use this API to delete a specific location from database",
        "operationId": "deleteLocation",
        "responses": {
          "204": {
            "description": "Successful delete operation. Location removed."
          },
          "404": {
            "description": "No location found with the given code."
          }
        },
        "tags": [
          "Location"
        ]
      }
    },
    "/v1/realtime": {
      "summary": "Operation for accessing realtime weather data based on client's IP address",
      "get": {
        "summary": "Returns the current weather information of the location based on client's IP address",
        "description": "Clients use this API to get real time weather data of client's location (based on IP address)",
        "operationId": "getRealtimeByIPAddress",
        "responses": {
          "200": {
            "description": "successful retrieval operation. Realtime data available for the client's location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": {
                      "type": "string",
                      "description": "name of location, in form of \"city name, state/region name, country name\"",
                      "example": "New York City, New York, United States of America"
                    },
                    "temperature": {
                      "type": "number",
                      "format": "int32",
                      "description": "current temperature of the location, default in Celsius degree",
                      "example": 12
                    },
                    "humidity": {
                      "type": "number",
                      "format": "int32",
                      "description": "percentage of humidity",
                      "example": 72
                    },
                    "precipitation": {
                      "type": "number",
                      "format": "int32",
                      "description": "percentage chance of rain",
                      "example": 50
                    },
                    "status": {
                      "type": "string",
                      "description": "current weather status",
                      "example": "Sunny"
                    },
                    "wind_speed": {
                      "type": "number",
                      "format": "int32",
                      "description": "current wind speed (default unit is km/h)",
                      "example": 8
                    },
                    "last_updated": {
                      "type": "string",
                      "format": "date-time",
                      "description": "the last time weather data is updated"
                    }
                  }
                },
                "examples": {
                  "realtimeWeatherExample": {
                    "summary": "realtime weather information",
                    "value": "{\n\"location\": \"San Francisco, California, United States of America\",\n\"temperature\": 12,\n\"humidity\": 32,\n\"precipitation\": 88,\n\"status\": \"Cloudy\",\n\"win_speed\": 5,\n\"last_updated\": \"2023-02-02 19:23:00\"\n}"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad request. Could not determine client's IP address"
          },
          "404": {
            "description": "No managed location found for the client's IP address"
          }
        },
        "tags": [
          "Realtime Weather"
        ]
      }
    },
...
    "/v1/hourly": {
      "summary": "Operation for accessing hourly weather forecast based on client's IP address",
      "description": "Clients use this API to get hourly forecast information for today",
      "get": {
        "summary": "Returns weather forecast for upcoming hours in the current day, based on client's IP address",
        "description": "Clients use this API to get weather forecast for upcoming hours in the current day. Location is determined automatically based on client's IP address.",
        "operationId": "listHourlyForecastByIPAddress",
        "responses": {
          "200": {
            "description": "successful retrieval operation. A JSON object representing hourly forecast information available for the client's location",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "location": {
                      "type": "string",
                      "description": "name of location, in form of \"city name, state/region name, country name\"",
                      "example": "New York, New York, US"
                    },
                    "hourly_forecast": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "hour_of_day": {
                            "type": "number",
                            "description": "the forecasted hour (in 24h format)",
                            "example": 14
                          },
                          "temperature": {
                            "type": "number",
                            "format": "int32",
                            "description": "forecasted temperature at the hour, default in Celcius degree",
                            "example": 19
                          },
                          "precipitation": {
                            "type": "number",
                            "format": "int32",
                            "description": "percentage chance of rain",
                            "example": 67
                          },
                          "status": {
                            "type": "string",
                            "description": "current weather status",
                            "example": "Cloudy"
                          }
                        }
                      }
                    }
                  }
                },
                "examples": {
                  "hourlyWeatherExample": {
                    "summary": "hourly weather forecast information",
                    "value": "{\n\"location\": \"San Francisco, California, United States of America\",\n\"hourly_forecast\":\n[\n{\n\"hour_of_day\": 8,            \n\"temperature\": 12,\n\"precipitation\": 88,\n\"status\": \"Cloudy\"\n},\n{\n\"hour_of_day\": 9,\n\"temperature\": 13,\n\"precipitation\": 86,\n\"status\": \"Cloudy\"\n},\n{\n\"hour_of_day\": 10,\n\"temperature\": 15,\n\"precipitation\": 60,\n\"status\": \"Sunny\"\n}                \n]\n\n}"
                  }
                }
              }
            }
          },
          "400": {
            "description": "bad request. Could not determine client's IP address"
          },
          "404": {
            "description": "No managed location found for the client's IP address"
          }
        },
        "tags": [
          "Hourly Forecast"
        ]
      }
    },
...
  },
...
}